+++ packages/expo-camera/android/src/main/java/expo/modules/camera/records/CameraRecords.kt
@@ -50,12 +50,14 @@ enum class VideoQuality(val value: String) : Enumerable {
 enum class FlashMode(val value: String) : Enumerable {
   AUTO("auto"),
   ON("on"),
-  OFF("off");
+  OFF("off"),
+  SCREEN("screen");
 
   fun mapToLens() = when (this) {
     AUTO -> ImageCapture.FLASH_MODE_AUTO
     OFF -> ImageCapture.FLASH_MODE_OFF
     ON -> ImageCapture.FLASH_MODE_ON
+    SCREEN -> ImageCapture.FLASH_MODE_SCREEN
   }
 }